Mod: GetSpotDetailMenu API request 수정, response 변수 수정

+++ data/src/main/kotlin/com/acon/data/api/remote/SpotApi.kt
@@ -23,6 +23,6 @@ interface SpotApi {
 
     @GET("/api/v1/spot/{spotId}/menus")
     suspend fun getSpotMenuList(
-        @Path ("spotId") spotId: Long
+        @Path ("spotId") spotId: Int
     ): SpotDetailMenuListResponse
 }
